[DEV-5856] continued file splitting
